APIsnmovies/requirements.txt
requests
python-dotenv
pandas
pytest
joblib
scikit-learn